Move Update to its own file, along with Event, so that they don't depend on Api.
Make is_pending an optional attribute defaulting to False
